[imagemath]: moved functions to image_processing.py
